iOS: tweak types to avoid warnings in 64 bit builds
Still more warnings left.

+++ platform/ios/Classes/MuChoiceFieldController.m
@@ -10,7 +10,7 @@
 {
 	void (^okayBlock)(NSArray *);
 	NSArray *choices;
-	int selected;
+	NSInteger selected;
 }
 
 - (id)initWithChoices:(NSArray *)_choices okayAction:(void (^)(NSArray *))block
